% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiverseTraceWorkshop-package.R
\docType{package}
\name{EpiverseTraceWorkshop-package}
\alias{EpiverseTraceWorkshop}
\alias{EpiverseTraceWorkshop-package}
\title{EpiverseTraceWorkshop: The Title of the Project}
\description{
A paragraph providing a full description of the project (on several lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/emiel200/EpiverseTraceWorkshop}
  \item Report bugs at \url{https://github.com/emiel200/EpiverseTraceWorkshop/issues}
}

}
\author{
\strong{Maintainer}: Em Prestige \email{emiel200@gmail.com} (\href{https://orcid.org/{{orcid}}}{ORCID}) [copyright holder]

}
\keyword{internal}
